requirements.txt
discord.py==2.6.4
discordLevelingSystem==1.2.1
discordLevelingSystem==1.2.1
python-dotenv==1.2.1
